fixed modifiers

+++ TaxiManagement0/BakeryProd.cs
@@ -1,3 +1,3 @@
 namespace TaxiManagement0;
 
-public abstract class BakeryProd(float price, float cost, float exDate) : Product(price, cost, exDate);
+public abstract class BakeryProd(float price, float cost, int exDate) : Product(price, cost, exDate);
